requirements.txt
Flask==2.3.2
tensorflow
scikit-learn
opencv-python-headless
joblib
numpy
torch
torchvision
Pillow
google-generativeai
python-dotenv
